Changed plugin.txt and telnet_reader
