+++ honeypot/plugins.txt
@@ -1,5 +1,5 @@
 http_reader
 http_reader2
-telnet_reader #comment
+telnet_reader
 https_reader
 ssh_plugin
